requirements.txt
duckdb>=0.9.2
pandas>=2.0.0
graphviz>=0.20.1
python-dotenv>=1.0.0
langchain>=0.1.0
streamlit>=1.30.0
chromadb>=0.4.22
langchain-community>=0.0.10
zowe-python-sdk-bundle